requirements.txt
Django~=4.2.7
Djangorestframework